fix in comments xml
